A spectral method for nonlocal diffusion operators on the sphere
This adds almost all of the code required to generate the figures and
data of the so-titled preprint with @Hadrien-Montanelli and Qiang Du:

https://arxiv.org/abs/1801.04902

As this repository grows, it’s anticipated that the code will take on
more of an OOP-style consistent with the rest of JuliaApproximation. So
for now (and with highly optimized fixed temporal and spatial
discretizations), this code could be considered for benchmarking
purposes.